LayoutTests/fast/forms/node-list-remove-button-from-form-expected.txt

Tests that removing a <button> from the document removes it from HTMLFormElement.elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Before removal of button:
PASS document.getElementById('aForm').elements is non-null.
PASS document.getElementById('aForm')['aButton'] is document.getElementById('aButton')

After removal of button:
PASS document.getElementById('aForm')['aButton'] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
